Add RPC caller and endpoint for FF-A
Adds RPC caller and endpoint using FF-A messaging: using direct requests
and shared memory buffers for communication.

The endpoint is only for running in SPs. The caller is implemented in
two flavours: for Linux and for SPs. The Linux version relies on the
FF-A user space interface described in external/ffa_tool/buf_common.h

Change-Id: I3b51bcaf114d06bbf3dad2d63529fa59d8977e19
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>