Fix: CROSS_COMPILE env variable is ignored
The "arm-linux" deployments ignore the CROSS_COMPILE environment
variable which makes the build fail when the compiler is not on the
PATH.
This patch fixes this incorrect behavior and makes the "arm-linux" and
the "opteesp" deployments aligned on how the environment variable is
handled.
The policy is to prefer any value already in the cache. If the variable
is not in the cache, the value will be taken over from the
environment.

Change-Id: Id3cc4d6c23767d8ad9cb9657a582d6f70e191e37
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>