Add LazyFetch
LazyFetch implements "lazy" external dependency management. In this
context lazy means, stop at the first available stage of the
dependency and ignore if it is not the latest version. This behavior
is optimized for build speed.
In addition, the module adds flexibility by allowing the build
environment to take control and use pre-built or pre-fetched content.

For more details, please see the inline documentation.

Signed-off-by: Imre Kis imre.kis@arm.com
Signed-off-by: Gyorgy Szing gyory.szing@arm.com
Signed-off-by: Benedek Tomasik benedek.tomasik@arm.com
Change-Id: I8522c41c5edfdad98048e7bc358965883d014a12