Add check of DebugFS driver API version
Currently the components in TS relying on the DebugFS PoC Linux driver
don't check which version of the driver is actually loaded. This can
lead to errors if the API version used by TS and the loaded module are
different. This commit introduces a runtime compatibility check to have
an informative error message and exit in case of incompatible APIs.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I6e8ea4143ac08b106e60516461611004c0b0b8c2