Add event log parameter handling in SP
A generic publish/subscribe mechanism is implemented to allow
boot information such as the TPM event log to be passed to
specific SPs, based on subscriptions specified in the SP
manifest.  This change implements the in-SP handling
for the initialization parameters.  The env_test has
been extended to include tests that check that the
event log is being passed and stored in the config store.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I89211ebffd9e2d1768eaf5736e6b28dceafb0a02