Peg to psa_arch_tests version with INHERIT toolchain support
Building with the psa_arch_tests external component previously
required a patch to be applied that allowed the toolchain
specified by the cmake toolchain file to be used to build the
psa_arch_tests library.  These changes have now been upstreamed
so the patch has been removed and the refspec updated.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I5fffb9aea2c00cc8e251e91e8d2bac554e4fbd1f