Add service discovery provider
Adds a partial implementation of the discovery provider to
enable clients to discover information about a service instance.
An instance of the discovery provider will be integrated into
each service provider to implement common operations for
information discovery.  This is initially needed to discover
the maximum request length to support multi-step crypto
operations.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Ib068a02ba9e77f8b893e1da9f1e36edb48ced182