Establish pattern for extending crypto service tests
To match the modular extensions to the crypto service provider,
this change establishes a pattern for extending service level
tests in a similarly modular way.  By componentizing groups
of test cases, a subset of tests may be used to match the
service provider under test.  The hash crypto extension is
used to establish the pattern.  Note that because the Cpp
crypto clients now use the same crypto caller as the PSA
Crypto client, the explicit PSA Crypto API client is no
longer needed and has been removed.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Ic8f975c9061f8aa0377dd59614862a4745a8c38c