Extend core crypto service level tests
Tests added for copy_key, purge_key and get_key_attributes
operations that have been recently added.  The copy_key
test identified a bug in the client-side translation
of key attributes retrieved from a service provider
where the read key lifetime was always persistent.
This is fixed in this commit.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I0ab7ca8c961c70a38627198c9a7556c34ce09dc9