Align attestation API test behaviour between deployments
Adds a missing claim source so that the set of claims is aligned
between the attestation/opteesp deployment and the PC standalone
service.  Also adds a missing psa crypto initalisation to the
psa-api-test/initial_attestation deployment.  Test behaviour
is now aligned when psa api tests are run in the different
environments.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I3529b0f5a7c9ecd2a62c32221db019f8423bd60c