Fix service provider chain test case
A test case to check chaining of service providers was building
a chain of providers where each array of handlers was of zero
length.  This was leading to a test failure when running on
an embedded platform but not natively on a PC.  This change
corrects the handler array sizes - test now passes on both
embedded and PC platforms.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: If8e2bd1c9322ae034a79ba79b158ee9f5d3d2944