Add newlib as external component
Add component for fetching and compiling newlib into a minimalistic
position independent library. It was necessary to introduce a patch
file for allowing aarch64-linux-gnu-gcc to compile newlib which is the
currently used compiler. Otherwise newlib would require a separate
aarch64-none-elf toolchain to be able to recognize the bare metal
target.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I12677bb89c4c46b81a29de8c087e7967908b8958