Fix: psa-api-test executable names clash
All psa-api-test executables used the same name which resulted in
executables being overwritten when installing to the same location.
As a result, b-test was not running all tests.
This commit changes the CMake project name for psa-arch-test
deployments, and the uses the PROJECT_NAME variable to create and
configure the built executable.

Change-Id: Idd80baf1cf7a9b16ad27b0d94abdb71c3179b169
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>