Enhance newlib fetch process
Update management of newlib external component to be optimized
for download speed insted of availability.
The updated process is:
  - check if binary is available. If yes configure build to use it
    and stop.
  - if not, check is source is available. If yes, build it and use
    the resulting binary.
  - if not, then download the source using git, compile it and use
    the resulting binary

The following variables can be set on the command line to alter the
behavior of the module:
  - NEWLIB_URL git repo URL to fetch from.
  - NEWLIB_REFSPEC revision to fetch
  - NEWLIB_SOURCE_DIR to specify location of source code in
    local file syetem.
  - NEWLIB_INSTALL_DIR to specify location of binary.

I.e. cmake -S <...> -B <...> -DNEWLIB_INSTALL_DIR=~/newlib
will make the resulting binary installed to ~/newlib. This can be
used later to speed up a clean build an use the prebuilt binary.

Change-Id: Idb4a39721f5739d25382d4749584981cb7e2d88f
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>