Fix: build failure of env-test
Since a365a04 the env-test deployment fails to build with the
following error:
  get_property could not find TARGET env_test.  Perhaps it has not yet
  been created.

The root cause is a spelling error in the CMakeLists.txt file of the
deployment. The error was most likely introduced by an incorrect
find-replace operation.

This change corrects the env-test deployment and makes it compile again.

Change-Id: I92ac7862d07f27956cda6f42e6413d7fbe7e3900
Signed-off-by: Gyorgy Szing Gyorgy.Szing@arm.com