Fix faulty SP build when build-id is enabled
Linkers are able to insert a build-id into the binary but by default it
is placed at the beginning of the binary. This conflicts the placement
of sp_head thus OP-TEE interprets the build-id as the sp_head structure
resulting in a load error.
The issue is fixed by placing the build-id section explicitly through
the linker script and adding an assert for checking sp_head placement.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I072bfba212e58b42dfc7b27f22af5bf90c8d2313