Allow smm_variable index UID to be overridden
The smm variable provider uses the backend NV store for
storing the variable index. A fixed UID is used to identify
it. For a storage backend that supports namespace partitioning
by owner, this should be fine. For backends that don't,
the UID is now build-time overridable to allow any clash
to be resolved.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Iaa723bf23860b6f769dc875163186411d0cd4b53