Fix: incorrect search path for newlib/libnosys.a
The cmake script is incorrectly using the NEWLIB_LIBC_PATH variable
to set NEWLIB_LIBNOSYS_PATH while removing NEWLIB_LIBNOSYS_PATH from
the cache.
As a result in some scenarios libnosys was not linked to SP binaries
and linking failed with errors like:
  stack_protector.c:(.text+0xc4): undefined reference to `_exit'

This change fixes the cmake script to use the correct variable.

Change-Id: I174221c35e23e639006e0d4c4b2ae739bc5e0b1b
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>