Simplify handling of psa_hash_suspend/resume
The PSA Crypto API functions psa_hash_suspend and psa_hash_resume
never need to be supported by a service provider so the client-side
handling can be simplified. This change removes the mapping from
the client functions to the backend crypto call, allowing the
crypto caller support for the functions to be removed.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Ib26e36e1e4065ebb685fb9cbf7f9c3a727784259