Fix bogus free() hazard in service framework tests
In the service framework tests, the common teardown function
calls direct_caller_deinit to clean-up. In normal cases,
this will result in buffers used by the direct caller
being freed. If a test assertion results in a test exit
without calling direct_caller_init, there is the hazard
that free will be called with uninitialized buffer pointers.
This change removes this hazard by initializing the direct_caller
structure to zero in the common setup method.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I635490991c6d2e6c00d44053b77d6c38f2446a73