Fix gcc_get_lib_location()
Error checking in gcc_get_lib_location() is invalid. GCC will not
return an error if the specified file cannot be fund, but will return
the filename without PATH. gcc_get_lib_location() was not acting
according to the above behavior and failures were silently ignored.
This change adapts the function to GCC's behavior and makes the
function correctly detect error conditions.

Change-Id: I5337cf01bc1190fa5b3106a3e93e4428ae7be236
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>