Fix libsp deployment
The libsp deployment is not using standard library specific settings of
the opteesp environment. As a result when the compiler uses a standard
library which has no sys/cdefs.h the build fails.
This patch adds newlib as an external dependency to the deployment and
ensures all mandatory build options are taken from the newlib
integration.

Change-Id: Ib47a359e3db7f705e31665c133b6f67c382a2813
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>