Fix use of GIT_SHALLOW in external components
External component cmake files are specifying GIT_SHALLOW TRUE for
the content fetch options for fetching upstream content. This
option is intended for cases where a remote branch is being
tracked and only the tip needs to be fetched. To maintain
control over the versions of external components used, a
specific tag or hash should be used as the refspec. For this
behaviour, GIT_SHALLOW TRUE cannot be used as fetches will
fail when the upstream repo changes. This change fixes the
problem.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I61cefb295352b0d2c46f6a426577605b6a3cfae6