Introduce uefi service component
In preparation for adding authenticated variable support, this
commit adds a service/uefi directory to act as a parent for
uefi related service components. The auth support that will
be added has broader scope than just uefi variables and may
also be used for capsule verification. Because of this,
it will live in its own directory.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Signed-off-by: Gabor Toth <gabor.toth2@arm.com>
Change-Id: Ie58f6537a8e105dfcf0904c8a80a27bdf219502a