Validate shared memory size in ffarpc_call_ep
In the FFA_CALL_OPCODE_SHARE_BUF call the size of the shared buffer
is forwarded in a register of an FF-A direct message. When the SP
retrieves the memory region this size parameter has to be checked
against the retrieved page count in order to prevent the SP from
overruning the buffer in case of a malformed request.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I1a91abc6687786fd87a10bcbc64a4b9235b046e6