Add storage partition authorizer
To allow for deployment specific authorization policy and lazy
resolution of a partition owner to a client ID, per-partition
access control can now use an authorizer function for cases where
no ACL has been configured during initialisation. This facility
may be used for SP deployments of the block storage service
provider to enable an FFA_GET_PARTITION_INFO to be performed
as a way to authorize a client on opening a partition.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I53ee3d3961fe5d2ce10bea8a692c87ae60ed1f1f