Fix: add missing parameter check to t_cose
t_cose uses PSA Crypto API for crypto operations and it needs the
headers of a PSA Crypto implementation. The PSA_CRYPTO_API_INCLUDE
variable can be used to set the location of the dependency and the
cmake script was not verifying if this input is set. This lead to
hard-to debug build errors when the parameter was missing.
This is fixed now with a check and appropriate error message.

Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>
Change-Id: Ifa7f7293c895109b4cc4c3e7482a900787dfc735