Make utils.h and compiler.h public headers
The mentioned two headers are used from a lot of components and
most library deployments depend on them. Make these public and be
part of the installed content.
Further fine-tuning might be needed as this change will make all
targets using the component installing the files even if not
actually needed.

Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>
Change-Id: I8e69f8eb20552d2bb53a740a8f478756fdce803e