Decouple SFS block size from underlying block store
The secure flash store limits the maximum file size to the block
size of the associated flash driver. In practice, the block size
of real flash devices will be too small to accommodate objects
such as RSA keys or certificates. To lift this restriction,
the sfs_flash_block_store_adapter can now present a larger block
size to the SFS and map read, write and erase operations to
multiple block store blocks. This allows the SFS to meet
storage requirements when working with real storage devices.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I113ae825d5373b67d9d2b32451d112cce047dfa3