Fix defects in remote test runner
While checking operation of the env_test SP, a couple of defects
in components built into the ts-remote-test deployment were observed.
The missing list tests functionality is added (-l option) and a regression
caused by a previous introduction of the 64-bit rpc_opstatus_t type that
meant that the list of test results wasn't printed is fixed.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I6ba92c939da167c12376709e75b6dbdbbb4f3a2b