Add tf-a external component
Allows code from tf-a to be compiled into TS deployments. tf-a includes
many components that can be reused in TS deployments, particularly in
the area of storage access and firmware update. This commit adds a
new tf-a external component which can be used to make tf-a components
available in TS builds. To provide a pattern for how tf-a components
can be used, a CFI NOR flash driver has been added. This shows how
the a component within the TS tree is used to reference source files
from the fetched tf-a repo. The CFI driver can be built but has not
actually been tested so more work is almost certainly needed to make
it usable. Even in its current state, the driver provides a reference
for how other components from tf-a can be reused. To enable a full
build to be tested that includes tf-a dependencies, the cfi-flash-optee
block-storage build configuartion has been added.

This allows
reuse of platform level drivers and code from TF-A. The first TF-A
code to reuse is a CFI NOR flash driver which is compatible with fvp
base platforms.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I69dacab1f4a0b2ab894e4130a0b54cff2a914563