Fix psa-iat-api-test build error
When a pre-built MbedTLS binary is used with the psa-iat-api-test, the
build fails with:
    Couldn't find psa/crypto.h

This is because MbedTLS.cmake is only defining PSA_CRYPTO_API_INCLUDE
when the source-code is fetched. This variable is used later to set an
include path which allows the compiler to find psa/crypto.h.
This change moves setting the variable to a ensure it is set always.

Change-Id: I63b9f796992ce03514438647ff8726026ce42cdc
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>