Introduce volume_index for compatability with tf-a
To allow for reuse of tf-a io_dev volume access components, the
tf-a interface function plat_get_image_source() needs to be
realized. The volume_index provides a realization where the
it acts as a singleton for holding per deployment volume ID
io_dev mappings. Deployment configuration code may add
mappings to suite the needs of the deployment. This helps to
isolate the tf-a io subsystem from tf-a specific configuration
conventions.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I913b6229d78be0c803ebe6a2f6cf06f8d1160c96