Add io device for block_store volume access
To enable tf-a components to access data from a block_store, an
io device is added that conforms to the tf-a model. The io device
presents a storage partition, accessed via a block_store, as a
single seekable volume. Any concrete block_store may be used with
this io device.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I3c8e9e198332d7c2601ccff03fa129ee164e5359