Support GPT formatted semihosted image file
Extends the block_store semihosting factory to include partition
configuration using the GPT in a prepared disk image file. This
change adds the capability for the semihosting block store to
calculate the size of storage from an existing disk image file.
This is needed to support the flow where a disk image is created
as part of the build and is accessed by the block storage SP to
allow for shared access.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I8898270c6b9c643d14784235bcee06743c97b212