Fix newlib build with musl-gcc
When newlib is built with yocto and musl-gcc newlib build system fails
to use gcc specific header files, which makes the build fail with an
error due to missing limits.h
This patch passes the include paths
discovered with GCC to the newlib build and fixes the issue.

In addition add explicit handling for CFLAGS and CFLAGS_TARGET
parameters of newlib .configure. NEWLIB_CFLAGS and NEWLIB_CFLAGS_TARGET
can be passed to cmake or be set in the environment to set these.

NEWLIB_EXTRA is renamed to NEWLIB_EXTRA_PARAMS and is extended to
support setting trough environment variable.

Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>
Change-Id: Ie5daa92fb510c5514587879bf3ac7075bca9788d