Extend storage partition info for FWU
The block_store interface provides a method for retrieving
information about storage partitions. This commit extends
the information retrieved to include the partition GUID
and the disk GUID of a containing parent, if it exists.
This extra information is needed to allow the FWU fw_directory
to be populated with information about where update images
need to be installed.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I57934a32feda540b2da5429d5333a4a597e3f308