Add fwu copy_installer
The copy_installer can be used to copy the contents from a source
volume into a destination volume. It is intended to be used to
handle the case where an incoming update package only describes
a partial update. Images not included in the update need to be
installed into the update bank using a copy of the previously
active bank.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Ibd2778b6efa5533adfb5fb309e2661e743d2dbff