Add direct_fw_inspector
The direct_fw_inspector can be used on platforms where direct
access to fw storage volumes is possible from the update agent
security domain. The direct_fw_inspector uses the set of
registered installers to provide the information needed to
populate the fw_director to provide a fresh view of updatable
images.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Iab6fe5d919a173a32c284dd5567e144692d58dba