Extend uuid component for FWU
This change extends uuid operations to better meet FWU and
GPT based configuration requirements. The change adds more
explicit support for the little-endian octet representation.
Although the little-endian octet order is not what's specified
in RFC4122, it is used by the UEFI specification which covers
disk and partition UUIDs. Because the acronym 'guid' is used
within the UEFI spec, the little-endian representation is
referred to as 'guid' in the names of uuid functions to
distinguish them from functions that assume standard octet
order.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Ie3d1f3d9fe44ad80c84e541c69fdd2cbfdacd128