Add SP binary UUID
When switching FF-A from service UUID to protocol UUID, all SPs will
use the same value. This means the FF-A UUID will not be suitable to be
used in file names as the names will collide.
As a solution this changes introduces "binary UUIDs" which identifies
the SP binary and can safely be used in file names.
The new SP_BIN_UUID_CANON argument of export_sp() can be used to set the
binary UUID. If not set, SP_UUID_CANON will be used as a fallback, and
thus the behavior is the same when not using protocol UUIDs.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>
Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I72562ca700759b6d0e2c8bbca0afee056af27719