Add power failure during update tests
Adds tests that simulate a power failure at different points during
the update flow. The tests are intended to check that the device is
never left in an un-bootable state during the update process when
unexpected power-cycles occur. Note that metadata atomic write
robustness is tested separately under the banked_fw_store tests.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I3326ab787509d208a3ff51d8381819d39929edf2